added sierpinsky triangle and kochs flake.

+++ Types.hpp
@@ -2,5 +2,5 @@
 
 #include <complex>
 
-typedef double real_t;
+typedef long double real_t;
 typedef std::complex <real_t> complex_t;
